backend/requirements.txt
fastapi
uvicorn
python-multipart
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
bcrypt
python-dotenv
pydantic
pypdf
langchain-google-genai
jobspy
pandas
pytest
